fix: Corregir feature tras fallo de comprensión del requisito
