+++ src/main/java/acme/realms/Consumer.java
@@ -12,14 +12,11 @@
 
 package acme.realms;
 
-import java.util.Objects;
-
 import javax.persistence.Entity;
 
 import acme.client.components.basis.AbstractRole;
+import acme.client.components.mappings.Automapped;
 import acme.client.components.validation.Mandatory;
-import acme.client.components.validation.Optional;
-import acme.client.components.validation.ValidNumber;
 import acme.client.components.validation.ValidString;
 import lombok.Getter;
 import lombok.Setter;
@@ -36,50 +33,14 @@ public class Consumer extends AbstractRole {
 	// Attributes -------------------------------------------------------------
 
 	@Mandatory
-	@ValidString(pattern = "^[A-Z]{2-3}\\d{6}$", message = "The identifier must follow the correct pattern")
-	private String				identifier;
-
-	@Mandatory
-	@ValidString(pattern = "^\\+?\\d{6,15}$", message = "The phone number must follow the correct pattern")
-	private String				phoneNumber;
+	@ValidString
+	@Automapped
+	private String				company;
 
 	@Mandatory
 	@ValidString
-	private String				physicalAddress;
-
-	@Mandatory
-	@ValidString(max = 50)
-	private String				city;
-
-	@Mandatory
-	@ValidString(max = 50)
-	private String				country;
-
-	@Optional
-	@ValidNumber(max = 500000)
-	private Integer				earnedPoints;
-
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = super.hashCode();
-		result = prime * result + Objects.hash(this.city, this.country, this.earnedPoints, this.identifier, this.phoneNumber, this.physicalAddress);
-		return result;
-	}
-
-	@Override
-	public boolean equals(final Object obj) {
-		if (this == obj)
-			return true;
-		if (!super.equals(obj))
-			return false;
-		if (this.getClass() != obj.getClass())
-			return false;
-		Consumer other = (Consumer) obj;
-		return Objects.equals(this.city, other.city) && Objects.equals(this.country, other.country) && Objects.equals(this.earnedPoints, other.earnedPoints) && Objects.equals(this.identifier, other.identifier)
-			&& Objects.equals(this.phoneNumber, other.phoneNumber) && Objects.equals(this.physicalAddress, other.physicalAddress);
-	}
+	@Automapped
+	private String				sector;
 
 	// Derived attributes -----------------------------------------------------
